Removed V1 versions of WeaponMaker
